chore: move `Core` part of python bindings into `Python/Core` (#4635)
This PR moves the python bindings of classes of `Core` from `Examples/Python` to `Python/Core/`.
It should not change any functionality.

95 % of this PR is shifting code, however, the clearer dependency structure is immediately visible by the remaining python bindings in the `Examples/Python` directory addressing only the "examples" module and only need the main module to attach submodules.

The latter mechanism will be changed in series of follow-up PRs.
