+++ CI/physmon/workflows/physmon_trackfinding_4muon_50vertices.py
@@ -161,7 +161,7 @@ with tempfile.TemporaryDirectory() as temp:
     )
 
     # Choosing a seeder only has an effect on VertexFinder.AMVF. For
-    # VertexFinder.IVF we always use acts.VertexSeedFinder.GaussianSeeder
+    # VertexFinder.IVF we always use acts.examples.VertexSeedFinder.GaussianSeeder
     # (Python binding is not implemented).
     # Setting useTime also only has an effect on VertexFinder.AMVF due to
     # the same reason.
@@ -182,7 +182,7 @@ with tempfile.TemporaryDirectory() as temp:
         trackParameters="trackParameters",
         outputProtoVertices="amvf_gauss_notime_protovertices",
         outputVertices="amvf_gauss_notime_fittedVertices",
-        seeder=acts.VertexSeedFinder.GaussianSeeder,
+        seeder=acts.examples.VertexSeedFinder.GaussianSeeder,
         useTime=False,  # Time seeding not implemented for the Gaussian seeder
         vertexFinder=VertexFinder.AMVF,
         outputDirRoot=tp / "amvf_gauss_notime",
@@ -195,7 +195,7 @@ with tempfile.TemporaryDirectory() as temp:
         trackParameters="trackParameters",
         outputProtoVertices="amvf_grid_time_protovertices",
         outputVertices="amvf_grid_time_fittedVertices",
-        seeder=acts.VertexSeedFinder.AdaptiveGridSeeder,
+        seeder=acts.examples.VertexSeedFinder.AdaptiveGridSeeder,
         useTime=True,
         vertexFinder=VertexFinder.AMVF,
         outputDirRoot=tp / "amvf_grid_time",
