+++ CI/physmon/workflows/physmon_trackfinding_ttbar_pu200.py
@@ -185,7 +185,7 @@ with tempfile.TemporaryDirectory() as temp:
         trackParameters="trackParameters",
         outputProtoVertices="amvf_gauss_notime_protovertices",
         outputVertices="amvf_gauss_notime_fittedVertices",
-        seeder=acts.VertexSeedFinder.GaussianSeeder,
+        seeder=acts.examples.VertexSeedFinder.GaussianSeeder,
         vertexFinder=VertexFinder.AMVF,
         outputDirRoot=tp / "amvf_gauss_notime",
         writeTrackInfo=True,
@@ -198,7 +198,7 @@ with tempfile.TemporaryDirectory() as temp:
         trackParameters="trackParameters",
         outputProtoVertices="amvf_grid_time_protovertices",
         outputVertices="amvf_grid_time_fittedVertices",
-        seeder=acts.VertexSeedFinder.AdaptiveGridSeeder,
+        seeder=acts.examples.VertexSeedFinder.AdaptiveGridSeeder,
         useTime=True,
         vertexFinder=VertexFinder.AMVF,
         outputDirRoot=tp / "amvf_grid_time",
