+++ Examples/Python/python/acts/examples/reconstruction.py
@@ -723,7 +723,7 @@ def addStandardSeeding(
     """
     logLevel = acts.examples.defaultLogging(sequence, logLevel)()
 
-    seedFinderConfig = acts.SeedFinderConfig(
+    seedFinderConfig = acts.examples.SeedFinderConfig(
         **acts.examples.defaultKWArgs(
             rMin=seedFinderConfigArg.r[0],
             rMax=seedFinderConfigArg.r[1],
@@ -968,7 +968,7 @@ def addOrthogonalSeeding(
     For parameters description see addSeeding
     """
     logLevel = acts.examples.defaultLogging(sequence, logLevel)()
-    seedFinderConfig = acts.SeedFinderOrthogonalConfig(
+    seedFinderConfig = acts.examples.SeedFinderOrthogonalConfig(
         **acts.examples.defaultKWArgs(
             rMin=seedFinderConfigArg.r[0],
             rMax=seedFinderConfigArg.r[1],
@@ -1077,7 +1077,7 @@ def addHashingSeeding(
     from acts.examples.hashing import SeedingAlgorithmHashing
 
     # Same configuration than the standard seeding
-    seedFinderConfig = acts.SeedFinderConfig(
+    seedFinderConfig = acts.examples.SeedFinderConfig(
         **acts.examples.defaultKWArgs(
             rMin=seedFinderConfigArg.r[0],
             rMax=seedFinderConfigArg.r[1],
@@ -2287,7 +2287,9 @@ def addVertexFitting(
     maxIterations: Optional[int] = None,
     initialVariances: Optional[List[float]] = None,
     useTime: Optional[bool] = False,
-    seeder: Optional[acts.VertexSeedFinder] = acts.VertexSeedFinder.GaussianSeeder,
+    seeder: Optional[
+        acts.examples.VertexSeedFinder
+    ] = acts.examples.VertexSeedFinder.GaussianSeeder,
     spatialBinExtent: Optional[float] = None,
     temporalBinExtent: Optional[float] = None,
     trackSelectorConfig: Optional[TrackSelectorConfig] = None,
