+++ Examples/Python/src/Alignment.cpp
@@ -25,7 +25,7 @@ using namespace ActsExamples;
 namespace ActsPython {
 
 void addAlignment(Context& ctx) {
-  auto [m, mex] = ctx.get("main", "examples");
+  auto& mex = ctx.get("examples");
 
   {
     auto ad =
