+++ Examples/Python/src/Detector.cpp
@@ -41,7 +41,7 @@ using namespace ActsExamples;
 namespace ActsPython {
 
 void addDetector(Context& ctx) {
-  auto [m, mex] = ctx.get("main", "examples");
+  auto& mex = ctx.get("examples");
 
   {
     py::class_<Detector, std::shared_ptr<Detector>>(mex, "DetectorBase")
