+++ Examples/Python/src/ExampleAlgorithms.cpp
@@ -7,7 +7,6 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 #include "Acts/Definitions/Algebra.hpp"
-#include "Acts/TrackFinding/TrackSelector.hpp"
 #include "ActsExamples/Fatras/FatrasSimulation.hpp"
 #include "ActsExamples/Io/Json/JsonGeometryList.hpp"
 #include "ActsExamples/Printers/ParticlesPrinter.hpp"
@@ -30,7 +29,7 @@ using namespace Acts;
 namespace ActsPython {
 
 void addExampleAlgorithms(Context& ctx) {
-  auto [m, mex] = ctx.get("main", "examples");
+  auto& mex = ctx.get("examples");
 
   mex.def("readJsonGeometryList", ActsExamples::readJsonGeometryList);
 
@@ -60,55 +59,6 @@ void addExampleAlgorithms(Context& ctx) {
     auto c = py::class_<Config>(alg, "Config").def(py::init<>());
 
     ACTS_PYTHON_STRUCT(c, inputTracks, outputTracks, selectorConfig);
-  }
-
-  {
-    using EtaBinnedConfig = TrackSelector::EtaBinnedConfig;
-    using Config = TrackSelector::Config;
-
-    auto tool = py::class_<TrackSelector>(m, "TrackSelector")
-                    .def(py::init<const Config&>(), py::arg("config"))
-                    .def(py::init<const EtaBinnedConfig&>(), py::arg("config"));
-
-    {
-      auto mc = py::class_<TrackSelector::MeasurementCounter>(
-                    tool, "MeasurementCounter")
-                    .def(py::init<>())
-                    .def("addCounter",
-                         &TrackSelector::MeasurementCounter::addCounter);
-    }
-
-    {
-      auto c = py::class_<Config>(tool, "Config").def(py::init<>());
-
-      patchKwargsConstructor(c);
-
-      ACTS_PYTHON_STRUCT(c, loc0Min, loc0Max, loc1Min, loc1Max, timeMin,
-                         timeMax, phiMin, phiMax, etaMin, etaMax, absEtaMin,
-                         absEtaMax, ptMin, ptMax, minMeasurements, maxHoles,
-                         maxOutliers, maxHolesAndOutliers, maxSharedHits,
-                         maxChi2, measurementCounter, requireReferenceSurface);
-
-      pythonRangeProperty(c, "loc0", &Config::loc0Min, &Config::loc0Max);
-      pythonRangeProperty(c, "loc1", &Config::loc1Min, &Config::loc1Max);
-      pythonRangeProperty(c, "time", &Config::timeMin, &Config::timeMax);
-      pythonRangeProperty(c, "phi", &Config::phiMin, &Config::phiMax);
-      pythonRangeProperty(c, "eta", &Config::etaMin, &Config::etaMax);
-      pythonRangeProperty(c, "absEta", &Config::absEtaMin, &Config::absEtaMax);
-      pythonRangeProperty(c, "pt", &Config::ptMin, &Config::ptMax);
-    }
-
-    {
-      auto c = py::class_<EtaBinnedConfig>(tool, "EtaBinnedConfig")
-                   .def(py::init<>())
-                   .def(py::init<const Config&>());
-
-      patchKwargsConstructor(c);
-
-      c.def_property_readonly("nEtaBins", &EtaBinnedConfig::nEtaBins);
-
-      ACTS_PYTHON_STRUCT(c, cutSets, absEtaEdges);
-    }
   }
 }
 }  // namespace ActsPython
