+++ Examples/Python/src/Hashing.cpp
@@ -31,7 +31,7 @@ void addHashing(Context& ctx) {
   auto hashingExampleModule = mex.def_submodule("_hashing");
 
   {
-    using Config = Acts::HashingAlgorithmConfig;
+    using Config = HashingAlgorithmConfig;
     auto c = py::class_<Config>(hashingModule, "HashingAlgorithmConfig")
                  .def(py::init<>());
     ACTS_PYTHON_STRUCT(c, bucketSize, zBins, phiBins);
@@ -39,7 +39,7 @@ void addHashing(Context& ctx) {
   }
 
   {
-    using Config = Acts::HashingTrainingConfig;
+    using Config = HashingTrainingConfig;
     auto c = py::class_<Config>(hashingModule, "HashingTrainingConfig")
                  .def(py::init<>());
     ACTS_PYTHON_STRUCT(c, annoySeed, f);
