+++ Examples/Python/src/Output.cpp
@@ -76,37 +76,7 @@ void register_csv_bfield_writer_binding(pybind11::class_<CsvBFieldWriter>& w) {
 namespace ActsPython {
 
 void addOutput(Context& ctx) {
-  auto [m, mex] = ctx.get("main", "examples");
-
-  ACTS_PYTHON_DECLARE_WRITER(ObjPropagationStepsWriter, mex,
-                             "ObjPropagationStepsWriter", collection, outputDir,
-                             outputScalor, outputPrecision);
-
-  ACTS_PYTHON_DECLARE_WRITER(
-      ObjSimHitWriter, mex, "ObjSimHitWriter", inputSimHits, outputDir,
-      outputStem, outputPrecision, drawConnections, momentumThreshold,
-      momentumThresholdTraj, nInterpolatedPoints, keepOriginalHits);
-
-  {
-    auto c = py::class_<ViewConfig>(m, "ViewConfig").def(py::init<>());
-
-    ACTS_PYTHON_STRUCT(c, visible, color, offset, lineThickness,
-                       surfaceThickness, quarterSegments, triangulate,
-                       outputName);
-
-    patchKwargsConstructor(c);
-
-    py::class_<Color>(m, "Color")
-        .def(py::init<>())
-        .def(py::init<int, int, int>())
-        .def(py::init<double, double, double>())
-        .def(py::init<std::string_view>())
-        .def_readonly("rgb", &Color::rgb);
-  }
-
-  py::class_<IVisualization3D>(m, "IVisualization3D")
-      .def("write", py::overload_cast<const std::filesystem::path&>(
-                        &IVisualization3D::write, py::const_));
+  auto& mex = ctx.get("examples");
 
   {
     using Writer = ObjTrackingGeometryWriter;
@@ -125,6 +95,15 @@ void addOutput(Context& ctx) {
                        containerView, volumeView, sensitiveView, passiveView,
                        gridView);
   }
+
+  ACTS_PYTHON_DECLARE_WRITER(ObjPropagationStepsWriter, mex,
+                             "ObjPropagationStepsWriter", collection, outputDir,
+                             outputScalor, outputPrecision);
+
+  ACTS_PYTHON_DECLARE_WRITER(
+      ObjSimHitWriter, mex, "ObjSimHitWriter", inputSimHits, outputDir,
+      outputStem, outputPrecision, drawConnections, momentumThreshold,
+      momentumThresholdTraj, nInterpolatedPoints, keepOriginalHits);
 
   ACTS_PYTHON_DECLARE_WRITER(CsvParticleWriter, mex, "CsvParticleWriter",
                              inputParticles, outputDir, outputStem,
