+++ Examples/Python/src/Propagation.cpp
@@ -67,42 +67,7 @@ void addPropagator(py::module_& m, const std::string& prefix) {
 
 namespace ActsPython {
 void addPropagation(Context& ctx) {
-  auto [m, prop, mex] = ctx.get("main", "propagation", "examples");
-
-  {
-    using Config = Navigator::Config;
-    auto nav =
-        py::class_<Navigator, std::shared_ptr<Navigator>>(m, "Navigator")
-            .def(py::init<>([](Config cfg,
-                               Logging::Level level = Logging::INFO) {
-                   return Navigator{cfg, getDefaultLogger("Navigator", level)};
-                 }),
-                 py::arg("cfg"), py::arg("level") = Logging::INFO);
-
-    auto c = py::class_<Config>(nav, "Config").def(py::init<>());
-
-    ACTS_PYTHON_STRUCT(c, resolveMaterial, resolvePassive, resolveSensitive,
-                       trackingGeometry);
-  }
-
-  {
-    using Config = Experimental::DetectorNavigator::Config;
-    auto nav =
-        py::class_<Experimental::DetectorNavigator,
-                   std::shared_ptr<Experimental::DetectorNavigator>>(
-            m, "DetectorNavigator")
-            .def(py::init<>(
-                     [](Config cfg, Logging::Level level = Logging::INFO) {
-                       return Experimental::DetectorNavigator{
-                           cfg, getDefaultLogger("DetectorNavigator", level)};
-                     }),
-                 py::arg("cfg"), py::arg("level") = Logging::INFO);
-
-    auto c = py::class_<Config>(nav, "Config").def(py::init<>());
-
-    ACTS_PYTHON_STRUCT(c, resolveMaterial, resolvePassive, resolveSensitive,
-                       detector);
-  }
+  auto [prop, mex] = ctx.get("propagation", "examples");
 
   ACTS_PYTHON_DECLARE_ALGORITHM(
       PropagationAlgorithm, mex, "PropagationAlgorithm", propagatorImpl,
@@ -113,54 +78,30 @@ void addPropagation(Context& ctx) {
   py::class_<PropagatorInterface, std::shared_ptr<PropagatorInterface>>(
       mex, "PropagatorInterface");
 
-  // Eigen based stepper
+  // Eigen stepper based propagator
   {
-    auto stepper = py::class_<EigenStepper<>>(m, "EigenStepper");
-    stepper.def(py::init<std::shared_ptr<const MagneticFieldProvider>>());
-
     addPropagator<EigenStepper<>, Navigator>(prop, "Eigen");
-  }
-
-  {
     addPropagator<EigenStepper<>, Experimental::DetectorNavigator>(
         prop, "EigenDetector");
   }
 
-  // ATLAS based stepper
+  // ATLAS stepper based propagator
   {
-    auto stepper = py::class_<AtlasStepper>(m, "AtlasStepper");
-    stepper.def(py::init<std::shared_ptr<const MagneticFieldProvider>>());
-
     addPropagator<AtlasStepper, Navigator>(prop, "Atlas");
-  }
-
-  {
     addPropagator<AtlasStepper, Experimental::DetectorNavigator>(
         prop, "AtlasDetector");
   }
 
-  // Sympy based stepper
+  // Sympy stepper based propagator
   {
-    auto stepper = py::class_<SympyStepper>(m, "SympyStepper");
-    stepper.def(py::init<std::shared_ptr<const MagneticFieldProvider>>());
-
     addPropagator<SympyStepper, Navigator>(prop, "Sympy");
-  }
-
-  {
     addPropagator<SympyStepper, Experimental::DetectorNavigator>(
         prop, "SympyDetector");
   }
 
   // Straight line stepper
   {
-    auto stepper = py::class_<StraightLineStepper>(m, "StraightLineStepper");
-    stepper.def(py::init<>());
-
     addPropagator<StraightLineStepper, Navigator>(prop, "StraightLine");
-  }
-
-  {
     addPropagator<StraightLineStepper, Experimental::DetectorNavigator>(
         prop, "StraightLineDetector");
   }
