+++ Python/CMakeLists.txt
@@ -1,1 +1,2 @@
+add_subdirectory(Core)
 add_subdirectory(Utilities)
