+++ Python/Core/setup.sh.in
