+++ Python/Core/src/GeometryGen1.cpp
@@ -28,9 +28,10 @@ using namespace pybind11::literals;
 using namespace Acts;
 
 namespace ActsPython {
-void addGeometryBuildingGen1(Context &ctx) {
-  auto m = ctx.get("main");
 
+/// This adds the geometry building bindings for the Gen1 geometry
+/// @param m the module to add the bindings to
+void addGeometryGen1(py::module_ &m) {
   using SurfacePtrVector = std::vector<std::shared_ptr<const Surface>>;
 
   py::class_<Layer, std::shared_ptr<Layer>>(m, "Layer");
