+++ Python/Core/src/MagneticField.cpp
@@ -6,11 +6,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "ActsExamples/MagneticField/MagneticField.hpp"
-
 #include "Acts/Definitions/Units.hpp"
 #include "Acts/MagneticField/BFieldMapUtils.hpp"
 #include "Acts/MagneticField/ConstantBField.hpp"
+#include "Acts/MagneticField/MagneticFieldContext.hpp"
 #include "Acts/MagneticField/MagneticFieldProvider.hpp"
 #include "Acts/MagneticField/MultiRangeBField.hpp"
 #include "Acts/MagneticField/NullBField.hpp"
@@ -57,8 +56,9 @@ Vector3 getField(MagneticFieldProvider& self, const Vector3& position,
 
 /// @brief Add the magnetic field bindings to a module.
 /// @param m the module to add the bindings to
-void addMagneticField(Context& ctx) {
-  auto& m = ctx.get("main");
+void addMagneticField(py::module_& m) {
+  py::class_<Acts::MagneticFieldContext>(m, "MagneticFieldContext")
+      .def(py::init<>());
 
   py::class_<MagneticFieldProvider, std::shared_ptr<MagneticFieldProvider>>(
       m, "MagneticFieldProvider")
