+++ Python/Core/src/Material.cpp
@@ -7,7 +7,6 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 #include "Acts/Geometry/GeometryContext.hpp"
-#include "Acts/MagneticField/MagneticFieldContext.hpp"
 #include "Acts/Material/BinnedSurfaceMaterialAccumulater.hpp"
 #include "Acts/Material/HomogeneousSurfaceMaterial.hpp"
 #include "Acts/Material/IMaterialDecorator.hpp"
@@ -23,11 +22,6 @@
 #include "Acts/Plugins/Json/ActsJson.hpp"
 #include "Acts/Plugins/Json/MaterialMapJsonConverter.hpp"
 #include "Acts/Utilities/Logger.hpp"
-#include "ActsExamples/Framework/ProcessCode.hpp"
-#include "ActsExamples/MaterialMapping/CoreMaterialMapping.hpp"
-#include "ActsExamples/MaterialMapping/MappingMaterialDecorator.hpp"
-#include "ActsExamples/MaterialMapping/MaterialMapping.hpp"
-#include "ActsExamples/MaterialMapping/MaterialValidation.hpp"
 #include "ActsPython/Utilities/Helpers.hpp"
 #include "ActsPython/Utilities/Macros.hpp"
 
@@ -44,20 +38,17 @@
 namespace Acts {
 class TrackingGeometry;
 }  // namespace Acts
-namespace ActsExamples {
-class IAlgorithm;
-}  // namespace ActsExamples
 
 namespace py = pybind11;
 using namespace pybind11::literals;
 
 using namespace Acts;
-using namespace ActsExamples;
 
 namespace ActsPython {
-void addMaterial(Context& ctx) {
-  auto [m, mex] = ctx.get("main", "examples");
 
+/// @brief Add the material bindings to a module.
+/// @param m the module to add the bindings to
+void addMaterial(py::module_& m) {
   {
     py::class_<ISurfaceMaterial, std::shared_ptr<ISurfaceMaterial>>(
         m, "ISurfaceMaterial")
@@ -84,38 +75,6 @@ void addMaterial(Context& ctx) {
         m, "IMaterialDecorator")
         .def("decorate", py::overload_cast<Surface&>(
                              &IMaterialDecorator::decorate, py::const_));
-  }
-
-  {
-    py::class_<MappingMaterialDecorator, IMaterialDecorator,
-               std::shared_ptr<MappingMaterialDecorator>>(
-        m, "MappingMaterialDecorator")
-        .def(py::init<const TrackingGeometry&, Logging::Level, bool, bool>(),
-             py::arg("tGeometry"), py::arg("level"),
-             py::arg("clearSurfaceMaterial") = true,
-             py::arg("clearVolumeMaterial") = true)
-        .def("binningMap", &MappingMaterialDecorator::binningMap)
-        .def("setBinningMap", &MappingMaterialDecorator::setBinningMap);
-  }
-
-  {
-    using Alg = MaterialMapping;
-
-    auto alg = py::class_<Alg, IAlgorithm, std::shared_ptr<Alg>>(
-                   mex, "MaterialMapping")
-                   .def(py::init<const Alg::Config&, Logging::Level>(),
-                        py::arg("config"), py::arg("level"))
-                   .def("scoringParameters", &Alg::scoringParameters)
-                   .def_property_readonly("config", &Alg::config);
-
-    auto c = py::class_<Alg::Config>(alg, "Config")
-                 .def(py::init<const GeometryContext&,
-                               const MagneticFieldContext&>());
-
-    ACTS_PYTHON_STRUCT(c, inputMaterialTracks, mappingMaterialCollection,
-                       materialSurfaceMapper, materialVolumeMapper,
-                       materialWriters, trackingGeometry, geoContext,
-                       magFieldContext);
   }
 
   {
@@ -229,21 +188,6 @@ void addMaterial(Context& ctx) {
   }
 
   {
-    auto mmca =
-        py::class_<CoreMaterialMapping, IAlgorithm,
-                   std::shared_ptr<CoreMaterialMapping>>(mex,
-                                                         "CoreMaterialMapping")
-            .def(py::init<const CoreMaterialMapping::Config&, Logging::Level>(),
-                 py::arg("config"), py::arg("level"));
-
-    auto c = py::class_<CoreMaterialMapping::Config>(mmca, "Config")
-                 .def(py::init<>());
-    ACTS_PYTHON_STRUCT(c, inputMaterialTracks, mappedMaterialTracks,
-                       unmappedMaterialTracks, materialMapper,
-                       materiaMaplWriters);
-  }
-
-  {
     auto mvc =
         py::class_<MaterialValidater, std::shared_ptr<MaterialValidater>>(
             m, "MaterialValidater")
@@ -258,23 +202,6 @@ void addMaterial(Context& ctx) {
     auto c =
         py::class_<MaterialValidater::Config>(mvc, "Config").def(py::init<>());
     ACTS_PYTHON_STRUCT(c, materialAssigner);
-  }
-
-  {
-    auto mv =
-        py::class_<MaterialValidation, IAlgorithm,
-                   std::shared_ptr<MaterialValidation>>(mex,
-                                                        "MaterialValidation")
-            .def(py::init<const MaterialValidation::Config&, Logging::Level>(),
-                 py::arg("config"), py::arg("level"))
-            .def("execute", &MaterialValidation::execute)
-            .def_property_readonly("config", &MaterialValidation::config);
-
-    auto c =
-        py::class_<MaterialValidation::Config>(mv, "Config").def(py::init<>());
-    ACTS_PYTHON_STRUCT(c, ntracks, startPosition, phiRange, etaRange,
-                       randomNumberSvc, materialValidater,
-                       outputMaterialTracks);
   }
 }
